+++ src/lib/PostSection.svelte
@@ -6,6 +6,7 @@ import Copy from '$lib/Copy.svelte'
 
 import CalendarIcon from 'svelte-bootstrap-icons/lib/Calendar.svelte'
 import PinAngleFillIcon from 'svelte-bootstrap-icons/lib/PinAngleFill.svelte'
+import PencilFillIcon from 'svelte-bootstrap-icons/lib/PencilFill.svelte'
 
 export let post
 </script>
@@ -14,14 +15,24 @@ export let post
   <section class="py2">
     <article>
       <Header>{post.title}</Header>
-      {#if post.date}
-        <aside class="px1 py1 md:px2">
-          <a class="h5 rlh4 inline-row items-center gap05" href="/{post.id}"
-            >{#if post.pinned}<PinAngleFillIcon />{:else}<CalendarIcon />{/if}
-            <span class="ul">{format(new Date(post.date), config.dateFormat)}</span></a
-          >
+
+      {#if post.date || post.author}
+        <aside class="px1 py075 md:px2 row-wrap gapx2 gapy025">
+          {#if post.date}
+            <a class="block py025 h5 rlh4 row gap05" href="/{post.id}"
+              ><span class="row py025"
+                >{#if post.pinned}<PinAngleFillIcon />{:else}<CalendarIcon />{/if}</span
+              ><span class="ul">{format(new Date(post.date), config.dateFormat)}</span></a
+            >
+          {/if}
+          {#if post.author}
+            <a class="block py025 h5 rlh4 row gap05" href="/archive/author/{post.authorId}"
+              ><span class="row py025"><PencilFillIcon /></span><span class="ul">{post.author}</span></a
+            >
+          {/if}
         </aside>
       {/if}
+
       <Copy>{@html post.content}</Copy>
     </article>
   </section>