+++ src/lib/slugify.js
@@ -1,0 +1,8 @@
+import slugify from 'slugify'
+
+export default function (x) {
+  if (typeof x !== 'string') return x
+  return slugify(x, {
+    lower: true,
+  })
+}